+++ Model/Alias.php
@@ -9,7 +9,7 @@ use Doctrine\ORM\Mapping as ORM;
 /**
  * Class Alias.
  *
- * @ORM\Entity
+ * @ORM\Entity(repositoryClass="RevisionTen\CMS\Model\AliasRepository")
  * @ORM\Table(name="alias")
  */
 class Alias
@@ -57,6 +57,19 @@ class Alias
      * @ORM\Column(type="text", nullable=true)
      */
     private $meta;
+
+    /**
+     * @var string
+     * @ORM\Column(type="string", nullable=true)
+     */
+    private $language;
+
+    /**
+     * @var Website
+     * @ORM\ManyToOne(targetEntity="Website", inversedBy="aliases")
+     * @ORM\JoinColumn(onDelete="CASCADE", nullable=true)
+     */
+    private $website;
 
     /**
      * @return string
@@ -210,6 +223,46 @@ class Alias
     public function setMeta(array $meta = null): self
     {
         $this->meta = json_encode($meta);
+
+        return $this;
+    }
+
+    /**
+     * @return string
+     */
+    public function getLanguage(): ?string
+    {
+        return $this->language;
+    }
+
+    /**
+     * @param string|null $language
+     *
+     * @return Alias
+     */
+    public function setLanguage(string $language = null): self
+    {
+        $this->language = $language;
+
+        return $this;
+    }
+
+    /**
+     * @return Website
+     */
+    public function getWebsite(): ?Website
+    {
+        return $this->website;
+    }
+
+    /**
+     * @param Website|null $website
+     *
+     * @return Alias
+     */
+    public function setWebsite(Website $website = null): self
+    {
+        $this->website = $website;
 
         return $this;
     }